swam-api/requirements.txt
flask 
request 
docker